server/requirements.txt
django
psycopg2-binary
djangorestframework
celery
django-celery-results
pillow
django-filter